поправил сущности, полность написал crud функционал

+++ backend/src/main/java/com/example/backend/repository/BICDirectoryEntryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BICDirectoryEntryRepository extends JpaRepository<BICDirectoryEntry, Integer> {
+public interface BICDirectoryEntryRepository extends JpaRepository<BICDirectoryEntry, Long> {
 }
