+++ backend/src/main/java/com/example/backend/repository/EmailMessageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmailMessageRepository extends JpaRepository<EmailMessage, Integer> {
+public interface EmailMessageRepository extends JpaRepository<EmailMessage, Long> {
 }
